mesh: consolidate placement tests into single tests.rs with category submodules
