+++ atom/mesh/src/core/placement/mod.rs
@@ -12,16 +12,4 @@ pub use types::{AdapterError, PlacementError, PlacementPlan, Protocol, RequestDe
 pub(crate) mod test_support;
 
 #[cfg(test)]
-mod tests_candidate;
-#[cfg(test)]
-mod tests_policy_apply;
-#[cfg(test)]
-mod tests_regular_planning;
-#[cfg(test)]
-mod tests_pd_planning;
-#[cfg(test)]
-mod tests_adapter;
-#[cfg(test)]
-mod tests_error;
-#[cfg(test)]
-mod tests_integration;
+mod tests;
